Moving hooks from JS -> TS
